Correción de errores

+++ models/movies.js
@@ -4,7 +4,7 @@ module.exports = (sequelize, type) => {
         date_created: type.DATE,
         image: type.STRING, // link
         score: {
-            type: type.INTEGER,
+            type: type.REAL,
             validate: {min: 1, max: 5}
         }
     })
